+++ apifox-models/src/main/java/cn/apifox/v1/models/api/item/ApiItem.java
@@ -9,7 +9,7 @@ import java.util.List;
 public class ApiItem {
     private String name;
 
-    private Long parentId;
+    private Integer parentId;
 
     private String serverId;
 
@@ -31,11 +31,11 @@ public class ApiItem {
         this.name = name;
     }
 
-    public Long getParentId() {
+    public Integer getParentId() {
         return parentId;
     }
 
-    public void setParentId(Long parentId) {
+    public void setParentId(Integer parentId) {
         this.parentId = parentId;
     }
